@@ -1,18 +1,16 @@
 package com.bridgefy.samples.fileshare;
 
 import android.Manifest;
-import android.bluetooth.BluetoothAdapter;
-import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.graphics.Color;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.LocalBroadcastManager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import androidx.core.app.ActivityCompat;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -21,6 +19,9 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.bridgefy.samples.fileshare.entities.Peer;
+import com.bridgefy.sdk.client.BFBleProfile;
+import com.bridgefy.sdk.client.BFEnergyProfile;
+import com.bridgefy.sdk.client.BFEngineProfile;
 import com.bridgefy.sdk.client.Bridgefy;
 import com.bridgefy.sdk.client.BridgefyClient;
 import com.bridgefy.sdk.client.Config;
@@ -30,12 +31,12 @@ import com.bridgefy.sdk.client.MessageListener;
 import com.bridgefy.sdk.client.RegistrationListener;
 import com.bridgefy.sdk.client.Session;
 import com.bridgefy.sdk.client.StateListener;
+import com.bridgefy.sdk.framework.exceptions.MessageException;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-
-import javax.security.auth.login.LoginException;
+import java.util.UUID;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -44,6 +45,8 @@ public class MainActivity extends AppCompatActivity {
     static final String INTENT_EXTRA_NAME = "peerName";
     static final String INTENT_EXTRA_UUID = "peerUuid";
     static final String INTENT_EXTRA_MSG  = "bridgefyFile";
+    static final String INTENT_BROADCAST_MSG  = "bridgefyFileProgress";
+    static final String INTENT_MSG_PROGRESS  = "bridgefyFileProgress";
 
     PeersRecyclerViewAdapter peersAdapter =
             new PeersRecyclerViewAdapter(new ArrayList<Peer>());
@@ -62,9 +65,7 @@ public class MainActivity extends AppCompatActivity {
         RecyclerView recyclerView = findViewById(R.id.peer_list);
         recyclerView.setAdapter(peersAdapter);
 
-
-
-
+        Bridgefy.debug = BuildConfig.DEBUG;
 
         Bridgefy.initialize(getApplicationContext(), new RegistrationListener() {
             @Override
@@ -101,18 +102,24 @@ public class MainActivity extends AppCompatActivity {
      */
     private void startBridgefy() {
         Config.Builder builder=new Config.Builder();
-        builder.setAntennaType(Config.Antenna.BLUETOOTH);
+        builder.setAntennaType(Config.Antenna.BLUETOOTH_LE);
+        builder.setAutoConnect(true);
+        builder.setEncryption(true);
+        builder.setMaxConnectionRetries(3);
+        builder.setBleProfile(BFBleProfile.BACKWARDS_COMPATIBLE);
+        builder.setEnergyProfile(BFEnergyProfile.HIGH_PERFORMANCE);
+        builder.setEngineProfile(BFEngineProfile.BFConfigProfileLongReach);
         Bridgefy.start(messageListener, stateListener,builder.build());
     }
 
 
 
-    private MessageListener messageListener = new MessageListener() {
+    private final MessageListener messageListener = new MessageListener() {
         @Override
         public void onMessageReceived(Message message) {
             Log.i(TAG, "onMessageReceived: ");
             // direct messages carrying a Device name represent device handshakes
-            if (message.getContent().get("device_name") != null) {
+            if (message!= null && message.getContent()  != null && message.getContent().get("device_name") != null) {
                 Peer peer = new Peer(message.getSenderId(),
                         (String) message.getContent().get("device_name"));
                 peer.setNearby(true);
@@ -128,6 +135,22 @@ public class MainActivity extends AppCompatActivity {
 
         }
 
+        @Override
+        public void onMessageDataProgress(UUID message, long progress, long fullSize) {
+            int currentProgress = (int) ((progress * 100) / fullSize);
+            LocalBroadcastManager.getInstance(getBaseContext()).sendBroadcast(
+                    new Intent(INTENT_BROADCAST_MSG)
+                    .putExtra(INTENT_MSG_PROGRESS, currentProgress)
+            );
+        }
+
+        @Override
+        public void onMessageFailed(Message message, MessageException e) {
+            Log.e(TAG, "Message sent failed: " + message.getUuid() + " sender: " + message.getSenderId());
+            LocalBroadcastManager.getInstance(getBaseContext()).sendBroadcast(
+                    new Intent(INTENT_BROADCAST_MSG).putExtra(INTENT_MSG_PROGRESS, 0)
+            );
+        }
     };
 
 
@@ -144,6 +167,19 @@ public class MainActivity extends AppCompatActivity {
         public void onDeviceLost(Device peer) {
             Log.w(TAG, "onDeviceLost: " + peer.getUserId());
             peersAdapter.removePeer(peer);
+            LocalBroadcastManager.getInstance(getBaseContext()).sendBroadcast(
+                    new Intent(INTENT_BROADCAST_MSG).putExtra(INTENT_MSG_PROGRESS, 0)
+            );
+        }
+
+        @Override
+        public void onDeviceDetected(Device device) {
+
+        }
+
+        @Override
+        public void onDeviceUnavailable(Device device) {
+
         }
 
         @Override
