@@ -7,11 +7,11 @@ import android.content.pm.PackageManager;
 import android.graphics.Typeface;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -66,6 +66,7 @@ public class MainActivity extends AppCompatActivity {
             bluetoothAdapter.enable();
 
             username = Build.MANUFACTURER + " " + Build.MODEL;
+            Bridgefy.debug = BuildConfig.DEBUG;
             // initialize the Bridgefy framework
             Bridgefy.initialize(getBaseContext(), registrationListener);
             setupList();
